+++ package.json
@@ -4,7 +4,7 @@
 	"scripts": {
 		"build": "lerna run build",
 		"release": "lerna run build && lerna publish --contents dist/release",
-		"prettier": "prettier --write  \"packages/**/{src,tests}/**/*.{ts,tsx}\"",
+		"prettier": "prettier --write  \"packages/**/{src,tests}/**/*.{ts,tsx,md}\"",
 		"create": "lerna create"
 	},
 	"devDependencies": {
