+++ packages/resources/src/provider.ts
@@ -38,7 +38,7 @@ export interface ResourceProviderProperties<S> {
 	renderer(resource: Resource<S>): RenderResult;
 }
 
-export function resourceProvider<S>(config: ResourceConfig<S>): Constructor<WidgetBase<ResourceProviderProperties<S>>> {
+export function provider<S>(config: ResourceConfig<S>): Constructor<WidgetBase<ResourceProviderProperties<S>>> {
 	const pathPrefix = uuid();
 	const { idKey = 'id' } = config;
 	const failedResourceProcess = createProcess(`${pathPrefix}-failed-resource`, [failedResource]);
@@ -117,4 +117,4 @@ export function resourceProvider<S>(config: ResourceConfig<S>): Constructor<Widg
 	return ResourceProvider;
 }
 
-export default resourceProvider;
+export default provider;
